refactor: move realtime out of codex-core
