+++ codex-rs/Cargo.toml
@@ -50,6 +50,7 @@ members = [
     "ollama",
     "process-hardening",
     "protocol",
+    "realtime",
     "realtime-webrtc",
     "rollout",
     "rmcp-client",
@@ -159,6 +160,7 @@ codex-plugin = { path = "plugin" }
 codex-model-provider = { path = "model-provider" }
 codex-process-hardening = { path = "process-hardening" }
 codex-protocol = { path = "protocol" }
+codex-realtime = { path = "realtime" }
 codex-realtime-webrtc = { path = "realtime-webrtc" }
 codex-responses-api-proxy = { path = "responses-api-proxy" }
 codex-response-debug-context = { path = "response-debug-context" }
