+++ codex-rs/realtime/BUILD.bazel
@@ -1,0 +1,7 @@
+load("//:defs.bzl", "codex_rust_crate")
+
+codex_rust_crate(
+    name = "realtime",
+    crate_name = "codex_realtime",
+    compile_data = glob(["templates/**/*.md"]),
+)
